[fix] export 特殊文字のエスケープを文字列表示時のみにするよう修正
上の修正の為、print_sorted_env.cの一部関数を、新ファイルsort_environ.cに分けました。
