+++ srcs/builtin/execute_export.c
@@ -14,34 +14,6 @@ static int	print_error(char *message)
 	return (EXIT_FAILURE);
 }
 
-static char	*set_value(char *line)
-{
-	char	*tmp;
-	char	*str;
-	int		i;
-
-	tmp = malloc(ft_strlen(line) * 2 + 1);
-	if (tmp == NULL)
-		return (NULL);
-	i = 0;
-	while (*line)
-	{
-		if (ft_strchr("\"$\'\\", *line))
-		{
-			tmp[i++] = '\\';
-			tmp[i++] = *line++;
-			if (line && ft_strchr("\"$\'", *line))
-				tmp[i++] = *line++;
-		}
-		else
-			tmp[i++] = *line++;
-	}
-	tmp[i] = '\0';
-	str = ft_strdup(tmp);
-	free(tmp);
-	return (str);
-}
-
 static bool	set_keyvalue(char *str, char **key, char **value)
 {
 	char	*equal;
@@ -56,7 +28,7 @@ static bool	set_keyvalue(char *str, char **key, char **value)
 		if (plus + 1 == equal && getenv(*key))
 			*value = ft_strjoin(getenv(*key), equal + 1);
 		else
-			*value = set_value(equal + 1);
+			*value = ft_strdup(equal + 1);
 	}
 	else
 		*key = ft_strdup(str);
