+++ srcs/utils/print_sorted_env.c
@@ -1,64 +1,45 @@
 #include "minishell.h"
 
-#define FRONT 0
-#define MID 1
-#define END 2
+static char	*create_escaped_str(char *str)
+{
+	char	*tmp;
+	char	*esc;
+	int		i;
+
+	tmp = malloc(ft_strlen(str) * 2 + 1);
+	if (tmp == NULL)
+		return (NULL);
+	i = 0;
+	while (*str)
+	{
+		if (ft_strchr("\"$\'\\", *str))
+		{
+			tmp[i++] = '\\';
+			tmp[i++] = *str++;
+			if (str && ft_strchr("\"$\'", *str))
+				tmp[i++] = *str++;
+		}
+		else
+			tmp[i++] = *str++;
+	}
+	tmp[i] = '\0';
+	esc = ft_strdup(tmp);
+	free(tmp);
+	return (esc);
+}
 
 static bool	get_str_literal(char *str, char **literal)
 {
-	char	*ptr;
+	char	*equal;
 
 	*literal = NULL;
-	ptr = ft_strchr(str, '=');
-	if (!ptr)
+	equal = ft_strchr(str, '=');
+	if (!equal)
 		return (false);
-	ptr = ft_str_sandwich(ptr + 1, "\"");
-	*literal = ft_strjoin("=", ptr);
-	free(ptr);
+	*literal = create_escaped_str(equal + 1);
+	*literal = ft_str_sandwich(*literal, "\"");
+	*literal = ft_strjoin("=", *literal);
 	return (true);
-}
-
-static void	merge(char **a, char **b, size_t index[3])
-{
-	size_t	i;
-	size_t	j;
-	size_t	k;
-
-	i = index[FRONT];
-	j = index[MID];
-	k = 0;
-	while (i < index[MID] && j < index[END])
-	{
-		if (ft_strncmp(a[i], a[j], INT_MAX) < 0)
-			b[k++] = a[i++];
-		else
-			b[k++] = a[j++];
-	}
-	if (i == index[MID])
-		while (j < index[END])
-			b[k++] = a[j++];
-	else
-		while (i < index[MID])
-			b[k++] = a[i++];
-	i = k;
-	while (i--)
-		a[index[FRONT] + i] = b[i];
-}
-
-void	sort_environ(char **a, char **b, size_t front, size_t end)
-{
-	size_t	mid;
-	size_t	index[3];
-
-	if (front == end || front + 1 == end)
-		return ;
-	mid = (front + end) / 2;
-	sort_environ(a, b, front, mid);
-	sort_environ(a, b, mid, end);
-	index[0] = front;
-	index[1] = mid;
-	index[2] = end;
-	merge(a, b, index);
 }
 
 char	**get_sorted_environ(void)
